// frontend/src/client/api/generated/model/enumOrganizationType.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Marumie Backend
 * OpenAPI spec version: 0.1.0
 */

export type EnumOrganizationType =
  (typeof EnumOrganizationType)[keyof typeof EnumOrganizationType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EnumOrganizationType = {
  household: "household",
  business: "business",
  nonprofit: "nonprofit",
  political_organization: "political_organization",
  other: "other",
} as const;