// frontend/src/client/api/generated/model/organizationRead.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Marumie Backend
 * OpenAPI spec version: 0.1.0
 */
import type { OrganizationReadDescription } from "./organizationReadDescription";
import type { EnumOrganizationType } from "./enumOrganizationType";
import type { OrganizationReadUserId } from "./organizationReadUserId";
import type { OrganizationReadSettings } from "./organizationReadSettings";

export interface OrganizationRead {
  id: string;
  name: string;
  display_name: string;
  description?: OrganizationReadDescription;
  type: EnumOrganizationType;
  slug: string;
  user_id?: OrganizationReadUserId;
  settings?: OrganizationReadSettings;
  created_at: string;
  updated_at: string;
}